updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -33,8 +33,9 @@ app.config['SECRET_KEY'] = 'SE3155'
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='pnguye55@uncc.edu').one()
-    return render_template('index.html', user=a_user)
+    if session.get('user'):
+        return render_template("index.html", user=session['user'])
+    return render_template('index.html')
 
 @app.route('/notes')
 def get_notes():
@@ -52,56 +53,60 @@ def get_note(note_id):
 
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        today = date.today()
-        today = today.strftime('%m-%d-%Y')
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            today = date.today()
+            today = today.strftime('%m-%d-%Y')
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='pnguye55@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    # check method used for request
-    if request.method == 'POST':
-        # get title data
-        title = request.form['title']
-        # get note data
-        text = request.form['noteText']
+    if session.get('user'):
+        # check method used for request
+        if request.method == 'POST':
+            # get title data
+            title = request.form['title']
+            # get note data
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
+            # update note data
+            note.title = title
+            note.text = text
+            # update note in DB
+            db.session.add(note)
+            db.session.commit()
+
+            return redirect(url_for('get_notes'))
+        else:
+            # GET request - show new note form to edit
+            # retrieve note from database
+            note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note=note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+@app.route('/notes/delete/<note_id>', methods=['POST'])
+def delete_note(note_id):
+    if session.get('user'):
+        # retrieve note from database
         note = db.session.query(Note).filter_by(id=note_id).one()
-        # update note data
-        note.title = title
-        note.text = text
-        # update note in DB
-        db.session.add(note)
+        db.session.delete(note)
         db.session.commit()
 
         return redirect(url_for('get_notes'))
     else:
-        # GET request - show new note form to edit
-        # retrieve user from database
-        a_user = db.session.query(User).filter_by(email='pnguye55@uncc.edu').one()
-
-        # retrieve note from database
-        note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=note, user=a_user)
-
-@app.route('/notes/delete/<note_id>', methods=['POST'])
-def delete_note(note_id):
-    # retrieve note from database
-    note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
-
+        return redirect(url_for('login'))
 @app.route('/register', methods=['POST', 'GET'])
 def register():
     form = RegisterForm()
